fixed compatibility issue with MSVC8
git-svn-id: https://supercollider.svn.sourceforge.net/svnroot/supercollider/trunk@5220 a380766d-ff14-0410-b294-a243070f3f08

+++ source/lang/LangSource/PyrLexer.cpp
@@ -164,7 +164,7 @@ double sc_strtof(const char *str, int n, int base)
 		else if (c >= 'A' && c <= 'A' + sc_min(36,base) - 11) z = z * base + c - 'A' + 10;
 		else if (c == '.') decptpos = i;
 	}
-	z = z / pow(base, n - decptpos);
+	z = z / pow((double)base, n - decptpos); 
 	return z;
 }
